Many improvements and fixes to the CPU activation test.
Skip the whole test if there is only one CPU.
On subtests that need to be repeated several times, use a number of repetitions that is linear to the number of CPUs.
Certain subtests can fail in some very unlikely situations and thus have been made weak.
Emit more diagnostics to the log about unexpected results.