Added "verbose" instrumenter that emits the instrumentation to stderr or to a file.
The NANOS_VERBOSE environment variable may contain a comma separated list of names that indicate which sections must be logged.
The matching is case insensitive. The "all" option enabkles all of them and is the default value. The rest can have "!" prepended
to them to disable them. For instace: NANOS_VERBODE="all,!LeaderThread".

The sections are:
	AddTask
	DependenciesByAccess
	DependenciesByAccessSequence
	DependenciesByGroup
	LeaderThread
	TaskExecution
	TaskStatus
	TaskWait
	ThreadManagement
	UserMutex

The NANOS_VERBOSE_TIMESTAMPS	environment variable indicates whether the logging should include timestamps. By default they are
recorded. To disable them set the environment variable to 0. In that case the order of the entries may not be reliable.

The NANOS_VERBOSE_FILE environment variable indicates a filename to store the log. By default the log is dumped to stderr.