Creating the infrastructure for Cluster support
Adding bits and pieces of the necessary infrastructure for running on a
cluster. At the moment we have defined what a message and a messenger
is and we have a basic MPI messenger implementation. I also added a
"hardware" representation of a cluster node and a basic cluster manager.

Cluster support is conditional and enabled at configuration time.

Also Cluster support is not enabled at runtime unless the environment
variable NANOS6_COMMUNICATION is set. The environment variable itself
determines the communication layer to be used. The only meaningful value
at the moment is 'mpi-2sided'