Force load balancing when threshold is 0
This is an issue in cases like heat. When there is a single task
creator, it may enqueue tasks locally. If, at some point, the tasks that
are being created are not pushed to the scheduler anymore (perhaps
because dependencies are not fulfilled), that scheduler node will not
rebalance, and enqueued tasks will not be sent to other CPUs.