Introducing virtual address space handling
We introduce a layer that pre-allocates and manages the virtual address
space that the runtime will manage both for internal and application
allocations.

We define a memory layout which takes into account the cluster memory
layout (if cluster support is enabled). We split the the address space
in distributed and local addresses. The distributed addresses are not
defined for the SMP case.

We split the local (intra-node) address space in one area per NUMA node.