Add dynamic private reduction storage management
- Private reduction storages were preallocated on 'ReductionInfo'
initialization and were fixed to the number of available CPUs. In this
commit this behaviour is changed in favour of a dynamic mechanism than
manages the number of required reduction private storages at each
moment. When such storages are unused, they are kept in a common storage
pool that allows them to be reused by other CPUs without having to
allocate/initialize a new storage.

- Reduction private storages and their metadata are now referred as
'Reduction slots'. References to 'ReductionCpu*' have been renamed
across the reduction components accordingly.

- Assert reduction tasks are executed in the host.