Make reduction storage allocation just before initialization
- As of now, allocation happens as soon as a free slot is requested.
In our structure, this forces us to allocate the storage while the lock
on the accesses is taken: The lock is taken during the process of
obtaining a free slot, as accesses need to be marked as they have used
that slot. Moving allocation in the function that actually returns the
initialized storage (given a slot ID), this can be done without the lock
taken.