A reduction is now combined by the last access within it
- The last access forming part of a reduction should combine the
reduction, as opposed for it to be combined by the first access after it
(previous behaviour).

- The motivation of this change is that it will hopefully allow
combining disjoint reductions whose data is required by the same access
in parallel. Before this, as both reductions were propagating to the
common access (who had to combine), its lock was taken and thus the
combinations were serialized.

- A reduction can be 'marked' as closed whenever a reduction access is
linked to:
    1) Another access that is not part of the same reduction
    2) A taskwait
    3) A top-level sink (this won't probably happen in correct code for now)

*Note for 1) - specially handle the case when the 'ReductionInfo'
allocation is caused by intersecting a missing region
(partially-overlapping reduction). In this situation, accesses have to
be marked *after* they have been processed.

- Avoid propagating read/write/concurrent satisfiabilities before a
reduction has been combined.

- Nesting (weakreductions) still not handled in this commit.