Add '_reductionTypeAndOperatorIndex' field in the bottom map
- This information is necessary when having nested (local) reductions of
a different type than the parent. See the following example:

    int x = 0;

    #pragma oss task weakreduction(+: x)
    {
        int tmp = 1;
        #pragma oss task reduction(*: tmp)
        {
            tmp *= 10;
        }

        // Closes local reduction *, but not +
        #pragma oss taskwait

        // Closes + reduction
        #pragma oss task reduction(+: x)
        {
            x += tmp;
        }
    }

    #pragma oss taskwait

- This is relevant because, when a reduction subaccess (in the bottom map)
is to be linked to a (nested) taskwait or top-level sink fragment, the
reduction access needs to be marked for closure (in order to be combined)
or not, depending on whether its '_reductionTypeAndOperatorIndex' matches
its parent's.
- Before this, only the 'accessType' was available, and no accesses were
marked for closure, not supporting local reductions of different types.