Added a mechanism for assigning unique ids to Messages
We created a ticket-like system to create Message identifiers. The ids
are 16-bit unsigned integers. This allows for up to 64K Messages per
node to be created throughout the execution of the application. We
assume that this is enough for the moment.

We use this id to create the MPI tags of the communications.