Update nanos6 API to receive dependency number in nanos6_create_task.
This change is directly tied to https://pm.bsc.es/gitlab/mercurium/mcxx/tree/nanos6-adding-total-num-deps
(Commits 60767f24, 763ecac3)

This change of API with Mercurium adds a size_t num_deps parameter to
the nanos6_create_task call, and is needed for the future implementation
of a discrete dependency system. Doing this change now is needed as well
to integrate the new dependency system into the automated performance
testing.