Differentiate workflow handling for 'wait' flag taskwait fragments
When a task is marked with the 'wait' flag (all Spawned functions are),
taskwait fragments are being created. This commit differentiates the
way we handle these fragments as compared to fragments related with a
normal taskwait operation. In short, we do not create a workflow for
handling the former, instead we mark them directly as completed.