Fixes in Execution Workflow
We clean-up the definition of DataLink and DataRelease steps. We also
re-order the Steps for the Workflow that executes a Task. Now the
notification Step is a successor to all the DataRelease Steps. This is
the correct thing to do, since the notification Step unregisters the
DataAccesses that are associated with the DataRelease Steps.