Keeping track of the MemoryPlace we run a task on
We track the MemoryPlace on which "run" the Task, by storing it within
the Task object, once we create the workflow for this Task. During the
unregister/release of the DataAccess objects of the Task, we use this
MemoryPlace to update their location, unless a valid (non-nullptr)
location is passed in the respective DataAccessRegistration call.

This should solve #45

Note: We setup the MemoryPlace for Taskloops in the Taskloop::run
method. We do that there, until we actually support Taskloops using the
execution workflow.