Changes in Messenger API
We removed the "send-to-many" call, because it seems quite cumbersome.
We also, changed the 'sendMessage' call to receive an optional parameter
to determine whether we will send the message in a blocking mode or not.
If we send it in non-blocking mode, the calling site needs to make sure
to use the new 'testMessageCompletion' API to check for Message
completion.