Changes in Messenger API for test completed messages
We now use `MPI_Testsome` instead of `MPI_Testany` for testing for
completed messages. Also, we extended the `Message` object to include a
flag to indicate whether or not it is delivered. This allows to simplify
the `testMessageCompletion` API by removing the `complete` Messages
parameter. The calling site can check whether the Message is delivered
calling directly the respective 'Message' method