Added a directory to track home nodes of memory regions
This commit introduces the concept of the home node of a memory region
and adds a mechanism to register and query the home nodes of memory
regions.

The point of a home node is to serve as the default MemoryPlace of the
said memory region, i.e. the MemoryPlace which holds the most updated
version of the memory region unless we have other information, e.g. a
location of a DataAccess.