Cleaning up DataLink and DataRelease steps
We do not store a pointer to DataAccess within DataLink and DataRelease
steps. Doing so, was dangerous since it is not safe to use a DataAccess
later on, since it might have changed. Instead, we store only the
relevant DataAccess fields we need for each type of step.

Also, we inline methods of the default DataLink and DataRelease steps.