Adding mechanism to check whether a DataReleaseStep should be triggered
We add a virtual DataReleaseStep method that checks if an associated
DataAccess is ready to release data. The method is meant to be
implemented from every ExecutionWorkflow implementation that uses the
DataReleaseStep, in order to define the conditions under which a data
release should be triggered (e.g. in cluster, CUDA, etc.)

Also, some style related fixes