Writing location information of taskwaits back to Task's accesses
Once a taskwait has been completed all location information it carried
is lost and we re-start the blocked Task with an empty BottomMap. As a
result, we do not know the location of sub-accesses that might be
created after the taskwait. In order to solve this, before we delete the
taskwait we return the location information it carried back to the
corresponding access.

In order to cover the case of local accesses, i.e. "accesses" that do
not appear in the dependency list of the Task, but are allocated withing
the context of the Task, we added support for adding these regions as
DataAccess objects in the list of accesseses of the Task that allocated
them. We remove that "fake" DataAccess once we de-allocate the region.