Splitting ClusterManager shutdown in remote and local operations
This commit splits the operation of shutting down in two distinct
operations. The first notifies all the ClusterNodes that we are shutting
down. The second one, performs all the local operations needed to
shutdown cluster.

The reason why this is needed is mainly Extrae. In Extrae mode, if we
try to shutdown instrumentation before the ClusterManager we enter in a
deadlock because Extrae is calling a MPI barrier which the rest of the
nodes will never call, because shutting down there has not started yet.