Register remote task before submitting it
There was a race condition found by @ksala, which could lead a remote
Task to be unregistered from the TaskOffloading mechanism before
registering it. This case could appear when offloading a Task without
any accesses. In this case, the task could be submitted, ran and
unregistered before running the TaskOffloading::registerRemoteTask.

This commit, fixes the issue by registering the Task before submitting
it. If there is satisfiability info to propagate, we do it by using the
TaskOffloading::propagateSatisfiability call, after submitting it.