Fixing race condition in remote satisfiability propagation
There is a race condition with registering remote tasks before submitting
them with the dependency system. If satisfiability information arrives
and gets handled after registering the remote task and before
submitting it to the dependency system it will get lost.

This commit, moves the handling of remote tasks in the TaskOffloading
system, where we serialize the process of handling new tasks and
respective asyncronous satisfiability info holding the lock that
registers the task, thus serializing correctly.