Fixing release of remote weak accesses
At the moment we update the location of accesses when we finalize them.
This should not happen for weak accesses, since they do not really
update the location of an access. However, when we offload a task with
a weak dependency, we do need to release the remote access even if it is
weak, because the remote task will probably create subtasks with strong
dependencies.

The problem comes when these strong dependencies do not cover the whole
region of the weak access. In that case, during release, we want to
update the location of the subregion covered by the strong access.

This commits modifies `unregisterTaskDataAccesses` to check whether the
access being finalized is weak. If so, it does not update its location.
In this way, we are sure that we update the location of a region only if
it has been touched by a strong access.