Increasing the number of Messages we can send
We use the Message id as part of the MPI tag we are sending. This limits
us to use 24 bits long ids. Up to now, we were using 16 bits. This
commit increases the limit to 24bits and also uses one counter per type
of Message, since the type is the other component we are using to
construct the MPI tag.