loader: Potential buffer overflow
The "n" argument of strncat indicates the maximum number of characters that can be appended to a string, not the total length of a string. On that ground, the strncat could actually overflow lib_name, and we were allowing MAX_LIB_PATH on each concat.

Substitute by snprintf, which is safer and maybe even clearer.