discrete: fix use-after-free in dependency release propagation
There is an issue in the discrete dependencies which will cause a use-after-free in certain situations.
The main thing to take into account is that we cannot access a task after calling decreaseDeletableCountOrDelete on it, because it might have been freed.
This patch solves this problem by delaying the calls to decreaseDeletableCount until we don't need the reference anymore.