CUDA: Replaced previous release's implementation; Introduced CUDADeviceInfo and StreamPool
From now on, HardwareInfo will initialize -if configured- a CUDADeviceInfo instance.
This will contain the Accelerator instances that actually represent the physical
devices present.

CUDAStreamPool contains preallocated cudaStreams to be used by the tasks, as well
as cudaEvents that we use for the task completion polling.

CUDAFunctions now serves as a helper class, wrapping various CUDA runtime calls
we need in other parts of the runtime.