Removed schedulers' dependency to the old device implementation; temporarily removed SubDeviceScheduler
As we removed all the previous DeviceInfoImplementation etc., scheduler code relying
on the way these classes used to be structured needed to be simplified and adjusted
to the new changes.

As SubDeviceScheduler also relied heavily on those and there is no actual
FPGA implementation merged yet, it was temporarily removed.