Fix external thread creation within CTF
This commit fixes three related issues:

1) An external thread might not have been initialized (it's ExternalThread
object does not exist yet) when it attempts to acquire the ctf external
threads shared buffer. This causes a nullptr dereference when trying to
obtain the external thread tid. This patch creates an ExternalThread
object when an uninitialized external thread tries to acquire it's
CPULocalData object.

2) The creation of an external thread triggers the
nanos6:external_thread_create tracepoint. This occurs when the external
thread attempts to emit any other tracepoint i.e. the external thread
tries to emit a tracepoints, it attempts to acquire it's per-cpu buffer,
when doing so it realizes that it's ExternalThread object does not exist
yet, it creates, it triggers the nanos6:external_thread_create
tracepoint, then it emits the tracepoint it was attempting to emit at
the beginning. However, the timestamp object of the initial event was
being acquired before the creation of the External thread object, which
means that events ended up unordered in the trace, which translates to
corrupted. This patch ensures that the timestamp is only acquired after
the external thread has been created.

3) The event timestamp in CTFAPI::tracepoint was being acquired outside
the stream lock. Under ExternalThread pressure, timestamps might not be
obtained in the order that events are written into the buffer, resulting
in a corrupted trace.