Fix new thread creation ctf race
The Instrument::createdThread() instrumentation point was used by CTF to
emit a traceopoint but at this point it is not guaranteed that the
current thread does owns its CPU object. This happens when a new thread
is created outside of the Nanos6 initialization. After the parent has
created the new thread, it still needs to emit at least the
"nanos6:thread_suspend" event, which was overlapping with the child's
"nanos6:thread_create" event and corrupting the buffer.