Annotate regions in the directory as requested by the user.
When using Nanos6 NUMA allocation, it may happen that the block size specified by the user is too small and so, the actual interleaving requested cannot be performed. In these cases, we want to annotate in the directory the homenode as if we were able to perform the requested interleaving.