Improving the configuration file, loader and Makefile.am
The configuration file has a version section where we provide the
options to chose the optimization level (disabled by default), the
dependency system (regions by default) and the instrumentation (none
by default) for the runtime library to load

From now on, we cleanly build the combination of all these three variant
types, except the instrumentations disabled by the user. Note there is no
graph variant for discrete dependencies for the moment

The "linear-regions-fragmented" value for the dependencies is no longer
valid when choosing the runtime variant. The only accepted name for the
region dependencies is "regions"