Devices: Fix call locations of setActiveDevice() calls
Placement of the calls can cause side effects in program execution.
This commit changes the setActiveDevice() calls to work-around strange
behaviors and broken contexts of multithreaded CUDA (that OpenACC also relies on)

CUDA:
 - only set the device when there are actual tasks to launch or check completion;
   previously it was set on every polling-service run.

OpenACC:
 - set the device also if going to processQueues(), but only if there are any active tasks.

This avoids setting the device during nanos6 bootstrap that the poling services are run.

More detailed explanation in issue #147