Adding NUMA-awareness.
When registering accesses, we try to propagate home node from parent or bottom map. If the access is new, then, we perform the move_pages method to ask the system which is the home node and set it into the data access.
In the scheduling part, we add a queue per NUMA node, and the scheduling decision is taken based on the ratio of data that belongs to each NUMA node.