Adapt taskfor address translation to compiler change
As part of a change to support a new feature, Mercurium and LLVM
will no longer overwrite the argsBlock with the result of translating
reduction addresses. Therefore, we have to remove a hack from the
taskfor that prevented translating addresses more than once.

Since programs built with an old compiler will no longer be correct,
update the loop API to reflect the breaking change.