Fix address translation for GPU tasks
We were previously translating the location of the argument, which was a
mistake, as we have to translate the argument itself.

This differs with the compiler-side implementation for address
translation, which prevents writing the argument itself by using the
stack instead and passing the new argument to the unpack function.

We could do that as well, but the approach used here should be fine, as
the translations are "undone" at a later point.