Setting number of spins before thread idle to one
Setting one as the number of spins that a thread must do before
being considered idle seems to have negligible performance penalty.
The resulting traces are around 7% bigger than when using 1024 spins
in an intensive benchmark. More info in !575

This commit maintains the templated function for busy waiting in the
delegation lock in case we need to change the number of spins in the
near future