Implementing API versioning and run-time checking
The API versioning from Mercurium is obsolete and is not implemented by our
LLVM-based compiler. We introduce a simpler API versioning that should replace
the older one. The new API is exposed in the nanos6/version.h header and
consists of the nanos6_check_version API function

The runtime supports a set of version families (initially a single one), where
each family has its major and minor version numbers. The compiler will add calls
to nanos6_check_version in the OmpSs-2 programs to check whether the required
versions are supported by the runtime. Unrecognized families or incompatible
major/minor versions must abort the execution in the runtime side