[CUDA] Fixing launching of CUDA task kernels
The ndrange clause in CUDA device tasks specify the global and local sizes of
thread hierarchies. The local size is the same as the CUDA block sizes. However,
the global size does not correspond to the CUDA grid size.

This issue is present since the refactor of the device API. The issue made the
runtime to launch many more CUDA blocks than the necessary. Thus, this problem
probably degraded the performance of our kernels.