Always pause task on enterTaskWait() for ovni
Regarding the ovni instrumentation, if0 tasks will require the parent
task to be paused before they begin the execution of a nested task due
to changes in ovniemu to make the task model more strict.

There are only two ways in which an if0 task can be executed:

1. Via If0Task::executeInline().

   With a call to the Nanos6 submit API and a ready if0 task. Reaches
   AddTask::submitTask() and directly calls If0Task::executeInline()
   because the task is ready to be executed. Which in turn, reaches the
   TrackingPoints::enterExecuteInline() and calls the instrumentation
   hook Instrument::enterTaskWait(). Later the inner task will be
   executed via currentThread->handleTask().

2. Via If0Task::waitForIf0Task() + If0Task::executeNonInline().

   With a call to the Nanos6 submit API and a non-ready if0 task, the
   if0 task will be registered in the dependency system and will return
   non-ready. The worker will then wait for the task to be completed by
   blocking itself with If0Task::waitForIf0Task(). This method will call
   its own tracking point and then Instrument::enterTaskWait() too,
   before blocking.

   When the task becomes ready, some other thread will put it in the
   scheduler for execution, and then it will be assigned via
   Scheduler::getReadyTask(). As it is an if0 task, that worker will
   call If0Task::executeNonInline(). Notice that the worker must not
   have a task assigned when is requesting another task to be executed,
   so there is no possibility that it runs as a nested task. When the
   if0 task finishes, the parent task will be added to the scheduler so
   the parent worker can be waken up. The parent then will finish by
   calling the Instrument::exitTaskWait() and resuming the parent task
   execution.

The only requirements to ensure that we don't emit two consecutive
executing events is to pause the parent task in If0Task::executeInline()
and in If0Task::waitForIf0Task(). Both methods call via the
corresponding tracking point to Instrument::enterTaskWait() with
fromUserCode set to false and then call Instrument::taskIsBlocked().

As in ovni we don't make use of the taskIsBlocked() family of events, we
simply block the task in the enterTaskWait() method. The fromUserCode
boolean was being used to determine if we should be pausing the task or
not. This was an incorrect usage. The task should always be paused, as
it won't be running user code anymore. The only usage of the
enterTaskWait() apart from If0Task is for TaskWait::taskWait(), in which
it makes sense to always pause the task.

This change won't affect much the visualization of task configurations
in Paraver for task wait events, as when the thread blocks with a
running task it won't be seen anyway as the thread will be Paused and it
will disappear from the timeline.