Fix thread instrumentation order in taskBlocks()
When a thread gets created in ThreadManager::getIdleThread() it may
begin the execution and switch to the CPU of the thread blocking. The
blocking thread was setting itself to the Cooling state too late, in the
TrackingPoints::threadWillSuspend().

By moving the tracking point before a new thread has a chance to be
created, there is no posibility that they can enter the same CPU in the
Running state.

Fixes: https://pm.bsc.es/gitlab/rarias/ovni/-/issues/165