Call ovni_thread_free() when threads end
The call causes the metadata to be written to disk and the stream buffer
to be destroyed. The stream will be marked as finalized. The call is
required to comply with the usage rules of libovni.

The stream is flushed always before the call, so we don't need another
call in pthreadEnd(). The instrumentation point Ovni::flush() is no
longer needed.