Reject non-instrumented threads in spawnFunction()
While the Nanos6 runtime accepts a call to the spawn function from any
context, this is not yet supported by the stream lifecycle model of
libovni: someone has to init the stream buffer and finalize it. To work
around this issue, a hack was introduced which was relying on a
property that was never guaranteed by libovni.

In particular, the first call to the function was initializing the
stream of the thread if it was not done before (threadMaybeInit) and
then flushing it to disk. Then, the stream was left initialized and no
call to ovni_thread_free() was ever done, but it was relying on
ovni_proc_fini() to eventually collect the stream buffer files.

This workaround won't work anymore, as the streams now require metadata
to be written by ovni_thread_free() to detect crashes. Also the behavior
of ovni_proc_fini() will change and it is likely it won't be able to
collect thread streams from the filesystem.

Therefore, we cannot support this case until we have support for
careless threads in libovni.

The change forces the thread to be instrumented by ovni prior to the
spawn call, and is responsibility of the caller to finalize it too. The
only case in which a non-instrumented thread was calling the spawn API
was used by TAMPI in a workaround that got removed. So as of now, AFAIK
all calls are done from a thread that is already instrumented.